+++ HISTORY.rst
@@ -17,3 +17,5 @@ History
 * LEICA-31: Run asset.update_metada() method every time afeter asset model instance change. (rudaporto)
 * LEICA-35: After Asset creation it will be automatic transitioned to pending state. (rudaporto)
 * LEICA-28: Improve models to import data from knack. (jsbueno) (rudaporto)
+* LEICA-36: Create events for Asset model instance lifecycle (POST, PUT, GET, DELETE) (rudaporto)
+